Add SimpleBancorp adapter
New Simple accounts will be created through BBVA. Preexisting accounts
were created through Bancorp and are slowly being migrated over.

It makes sense to have a legacy adapter for those still on Bancorp, but
the default adapter for Simple should be the BBVA adapter. This renames
the Simple adapter to SimpleBancorp (as my account was under Bancorp
when this was created).